Adding functions to label a scene
According to the documentation, the semantic segmentation camera
provides the labels of each object in the scene as a tag in the red
channel. The code extracts the red channel, maps the labels with their
respective values and saves everything under the DATA_OUT folder.

Three facts are noteworthy. First is that these functios should soon be
moved to a class that handles only aspects related to cameras. Second,
the dictionary that maps the values in the red channel with the tags was
extracted from the source code, and not from the documentation. The
latest version of the documentation ranges the values from 0 to 22, and
often we were getting higher values. By searching the source code from
the cpp carla's lib, we found the mapping that is performed behind the
scenes. This mapping was the one added.  Third, we had some trouble
when verifying if the labels assigned were correct. We haven't put much
effort to it, though.